requirements.txt
alabaster==0.7.12
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
bokeh==2.1.1
branca==0.4.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coverage==5.2
coveralls==2.1.1
cycler==0.10.0
datascience==0.15.6
decorator==4.4.2
docopt==0.6.2
docutils==0.16
folium==0.11.0
idna==2.10
imagesize==1.2.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
more-itertools==8.4.0
mysql-connector-python==8.0.21
numpy==1.19.0
packaging==20.4
pandas==1.0.5
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
prompt-toolkit==3.0.5
protobuf==3.12.2
py==1.9.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
scipy==1.5.1
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.9
wcwidth==0.2.5
wrapt==1.12.1
